requirements_calibration_data_agg.txt
apache-airflow==2.6.3
psycopg2-binary
yfinance
pylru==1.0.9
scipy
pyhocon==0.3.37
statsmodels
ta-lib==0.4.17
google-cloud-bigquery==2.13.1
google-cloud-storage==1.37.1
fredapi==0.4.0
pandas-datareader==0.10.0
pyarrow
tables
gcsfs
fsspec
google-auth
google-auth-oauthlib
google-auth-httplib2